requirements.txt
Django==1.7.4
Fabric==1.10.1
Jinja2==2.7.3
Markdown==2.5.2
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.7.0
beautifulsoup4==4.3.2
django-annoying==0.8.0
django-blog-zinnia==0.15.1
django-bootstrap-form==3.1
django-contrib-comments==1.5
django-debug-toolbar==1.2.2
django-markup-deprecated==0.0.3
django-mptt==0.6.1
django-sslify==0.2.5
django-tagging==0.3.4
django-xmlrpc==0.1.5
ecdsa==0.11
glue==0.9.4
paramiko==1.15.2
polib==1.0.6
psycopg2==2.5.4
pycrypto==2.6.1
pymongo==2.8
pyparsing==2.0.3
pytz==2014.10
requests==2.5.1
sorl-thumbnail==12.2
sqlparse==0.1.14
wsgiref==0.1.2
